[#24275] YSQL: Make connection sticky for base or shell type create type stmts with conn mgr
Summary:
This diff makes the connection sticky on executing create type query of the form `shell type`
 or `base type` with ysql connection manager.
The problem is sometime on the concurrent postgres backend processes the definition of create type defined in one session is not visible to other, due to catalog cache not getting bump up.
This is known issue in the YB, therefore as a workaround to support create type of the given above forms with connection manager, the session is made sticky.
Jira: DB-13167

Test Plan:
Jenkins: enable connection manager

Make sure below test is passing with ysql connection manager:
```-java-test org.yb.pgsql.TestPgRegressPgMisc#testPgRegressPgMisc```

Reviewers: skumar, stiwary, rbarigidad

Reviewed By: stiwary

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38506