[BACKPORT 2.20.7][PLAT-11305] Adding prometheus log file
Summary:
Original commit: 2d77f2313506bce0d6028608bfc38c60295f81b6 / D38855
We were not passing the prometheus log file anywhere to actually write the logs but were showing it in the status output which was kind of confusing. This diff adds the StandardOutput and StandardError redirection fields directly to the service file.

NOTE: This only works with systemd >= 236 which should cover most modern linux distros. If the version is < 240 we will use file mode which will replace the file on each restart of the service, but at least it's logged to a file in this case.

Test Plan: install yba and make sure prometheus logs present

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware, svc_phabricator, steve.varnau

Differential Revision: https://phorge.dev.yugabyte.com/D38943