[PLAT-15639]: YBA UI map shows continents instead of country borders
Summary:
Currently YBA UI shows incorrect maps for the India region. Let’s ensure that we show the correct maps.
At present, we are showing the borders of individual countries, which could potentially present challenges for the company. Given that we aim to serve customers globally, displaying national borders might lead to issues, as certain
countries have territorial
disputes or differing interpretations of their boundaries. To avoid potential conflicts and ensure neutrality, lets consider consider shifting our focus from country-specific borders to broader geographic regions or continents.

Steps to do it:
1. Download TileMill or build from source https://tilemill-project.github.io/tilemill/
https://tilemill-project.github.io/tilemill/docs/install/
2. Run it locally based on steps given in above link
3. Select "New Project" and give any file name and select add
4. Click the icon as shown in the image and delete the countries layout
{F297619}
5. Click "Add Layer" and under the Dialog, select File tab , select Data Source as "Map Box" as shown in images
{F297615}
6. Select "natural-earth-1.4.0" -> cultural -> 10m-admin-0-countries.zip and select "Done" and hit "Save"
7. Select Export -> MBTiles as shown in below image
{F297612}
8. Now select the range of map you want to show and give it a name and hit "Save" as shown in image below
{F297611}
9. Now .mbtile file will get generated
10. Follow the steps mentioned in README.md

Also in step 8, sometimes file may not get downloaded and can often error out, play around with zoom level and that might download it sooner
After uploading it to s3 bucket, (this is a cumbersome process) based on zoom level, there can be about 40-50 images and we need to manually change permissions individually. Tried to see if we can do it bulk, but did not allow it

Test Plan:
Please refer to screenshot
{F298129}

{F298130}

{F298131}

NOTE: This was added 7 years back and its(React-leaflet) an old library and no longer actively supported

Reviewers: jmak, kkannan

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38946