[#24394] docdb: Fix heap-use-after-free in TransactionParticipant::Impl::SubmitUpdateTransaction
Summary:
TransactionParticipant::Impl::SubmitUpdateTransaction moves the passed in
std::unique_ptr<tablet::UpdateTxnOperation> to TabletPeer::SubmitUpdateTransaction, but then
calls ToString on it in event of failure, but at this point the operation has been destroyed.

This diff changes this codepath to not destroy the operation object on failure.

This is the cause of YbAdminSnapshotScheduleTest.ConsistentTxnRestore failures.
Jira: DB-13303

Test Plan: Jenkins

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38953