[BACKPORT 2024.2][PLAT-15642][PLAT-15517]: Fix xCluster table status and bi-directional replication bug
Summary:
In this case, we have an extra table on the target(table_2) whose table ID matches the table's ID on the source universe, and in replication(table_1), the YBA was setting targetTableInfo incorrectly, i.e. target table info of the table_1.
Before even checking the table status, we used to assume that the table ID present in the table config is from the source universe, and from there we used to fetch the consumer table ID and set target table info but if the table is in status extra table on target, The table config ID is from target universe and we do not need to get the consumer table id, we can directly set the target table info.

Original diff/commit: 03c785a63c00c757ca6459795447ae799c826b17/D38956

Modified the previous xCluster bootstrap check to prevent users from adding a new table in bi-directional replication if it is missing from the target universe.

Original diff/commit: 5d0bd776c86ce00617ee23884de439af8f1aa5f2/D38866

Test Plan:
The issue occurred intermittently and could not reproduce consistently but was able to understand the issue from the code itself.

Local provider test

Reviewers: cwang, sanketh, hzare

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39011