[#18777] CDCSDK: Optimize reads by reading the WAL segment by segment in GetChanges
Summary:
In the current implementation (before the changes in this diff) we read the entire WAL starting from `from_op_id` for every GetChanges call. This is required because in every GetChanges call we want to stream all the records with commit time less than or equal to the consistent stream safe time (calculated at the time of GetChanges call). This can lead to significant disk read overhead in the case where WAL to be read is large.

In this diff, a new mechanism for reading the WAL has been introduced. A new method named `ReadReplicatedMessagesInSegmentForCDC` has been introduced. This method takes the from_op_id as its argument and finds the segment number in which the op corresponding to the from_op_id lies. If the `from_op_id` belongs to the currently active segment then in this case we fallback to the existing mechanism of reading the entire WAL. Otherwise we read all the ops from the segment starting to which the from_op_id belongs from the from_op_id. We also return the value of consistent stream safe time read from the footer.

In cdcsdk_producer in `GetConsistentWALRecords` we iterate over the ops returned by `ReadReplicatedMessagesInSegmentForCDC`and determine if there exists atleast one op with commit time less than or equal to consistent_stream_safe_time. If no valid records are found in the segment then we proceed to read the next segment and accordingly update the consistent_stream_safe_time. This process is repeated until atleast one valid record is found.
Jira: DB-7658

Test Plan:
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_stream-test --gtest_filter CDCSDKConsistentStreamTest.TestReadingOfWALSegmentBySegment
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_stream-test --gtest_filter CDCSDKConsistentStreamTest.TestReadingOfWALWithUncommittedTxn
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_stream-test --gtest_filter CDCSDKConsistentStreamTest.TestConsumptionContinuationAfterSegmentGC

Reviewers: skumar, asrinivasan, siddharth.shah, esheng, rthallam, sergei

Reviewed By: sergei

Subscribers: siddharth.shah, ybase, svc_phabricator, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37941