[PLAT-15703] Make YBC upgrade use Ansible to fix issues like RPC server unavailable
Summary:
Earlier we were using YBC rpc to upgrade YBC servers and the flow differed for K8s and VMs.
Now we will not be using RPC for VM upgrades too, and instead use ansible which is used by general
YBA tasks. The benefit of doing this is we can control it even in cases where YBC server is not
available.

It makes the flow consistent and more controllable

Test Plan: Manually verified YBC upgrade was working file. Fixed UTs to follow the new flow.

Reviewers: vpatibandla, vbansal

Reviewed By: vpatibandla

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39016