[#20908] YSQL: Gate inplace index updates behind feature enablement GUC
Summary:
D36588 introduced an optimization that enabled updates to non-key columns of secondary indexes to be performed in-place (ie. not require a DELETE + INSERT).
This optimization was ON by default and was not gated behind a feature flag.

This revision introduces a Postgres GUC `yb_enable_inplace_index_update` to enable/disable the feature.
By default the GUC is set to true, keeping the optimization turned ON.
Jira: DB-9891

Test Plan:
Run the following tests:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressUpdateOptimized#schedule'
```

Reviewers: smishra, amartsinchyk, tnayak

Reviewed By: smishra

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38763