[BACKPORT 2024.2][PLAT-15514] Allow enter key to create search tokens
Summary:
Pressing `enter` after inputting the searh term is natural for the user so we should accept `enter` keypress as a
trigger for creating a search token from the current input.

Original commit: D39084

Test Plan: Verify the user is able to use both `tab` and `enter` key press to create search tokens.

Reviewers: rmadhavan, kkannan, cwang, hzare

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39284