[BACKPORT 2024.1][#21836] docdb: Fix blind write in heartbeat processing for RBS' tablets.
Summary:
Original commit: rYBDB396433e3c109 / D34392
The heartbeat processing code in master relies on the consensus metadata for the reported tablets to ignore stale heartbeats. However when a TS is remote bootstrapping a new tablet, the reports it sends for this tablet do not have the consensus metadata field filled. When the master leader receives such a report it blindly overwrites its existing tablet replica state in memory. These blind overwrites can cause the master leader to corrupt its in memory metadata for a tablet, losing track of the leader. See the linked GH issue for more information on the scenario.

This diff uses the sequence number already included in tablet reports to ignore tablet reports from a TS that were sent before the most recently processed tablet report for a given tablet.
Jira: DB-10737

Test Plan:
```
% ybd --with-tests --cxx-test-filter-re 'master_heartbeat-itest' --cxx-test master_heartbeat-itest --gtest_filter "MasterHeartbeatITest.IgnoreEarlierHeartbeatFromSameTSProcess"

% ybd --with-tests --cxx-test-filter-re 'master_heartbeat-itest' --cxx-test master_heartbeat-itest --gtest_filter "MasterHeartbeatITest.ProcessHeartbeatAfterTSRestart"
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D39115