[PLAT-15657] Add error banner and tooltips for table status
Summary:
This diff adds a warning banner on the xCluster replication and xCluster DR config pages which calls to attention tables which are
added/dropped on one side of the replication but not the other.
DLL operations must be performed in a specific order which we have provided in the public docs:

When this order is not followed or the steps required are incomplete, the warning banner will be shown.

This diff also adds tooltips to the table statuses which pertain to the mismatched database schema.

Test Plan:
Verify that the xCluster replication and xCluster DR config pages displays a warning banner indicating database schema
mismatch when at least one of:
- Extra table on source
- Extra table on target
- Dropped table on source
- Dropped table on target
is present on the xCluster config.
Verify that a tooltip appears on hover for the above table statuses with the appropirate explaination text.
{F299059}
{F299060}
{F299061}
{F299062}

Reviewers: sanketh, cwang, vbansal, rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38976