[PLAT-15707] Updated ntpq parsing to assume milliseconds, not seconds
Summary:
ntpq -p offset returns the offset in milliseconds, not in seconds, updated the parsing as needed.

In addition, improved the awk check to look for the line starting with '*' as that is the active server.
If there is no active server, return error

Test Plan:
deployed a universe and used ntp. Validated health checks passed with no time drift and
failed with a drift

Reviewers: muthu, nsingh

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39150