[PLAT-15664] Non-restart GFlag upgrade is causing the audit log flags to be wiped off
Summary:
Non-restart GFlag upgrade is causing the audit log flags to be wiped off.
The rolling restart gflag upgrade works as expected, but the non-restart gflags are overriding the pgaudit settings when DB audit log is enabled. When a rolling restart is done, the DB audit logging is resumed again properly.

This diff sets the missing pgaudit log settings in the memory for a non-restart gflag upgrade.

Test Plan:
Manually tested:
1. Created a universe with DB audit logging enabled, changed a key value pair in the user set gflag `ysql_pg_conf_csv`, did a non-restart gflag upgrade. We now see both key value pairs (user set one and YBA generated one for DBAL), so it works successfully. (This flow failed before this change)
2.  Created a universe without DB audit logging enabled, changed a key value pair in the user set gflag `ysql_pg_conf_csv`, did a non-restart gflag upgrade, works successfully.

Apart from this,
Run UTs.
Run itests.

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39136