[PLAT-15759] Prometheus rollback may break because of symlinks
Summary: Prometheus rollback can break because we rely on symlinks from /usr/local/bin/ to the prometheus binary, which is a little unnecessary. This diff changes the symlink to be entirely contained within the software directory so that on rollback we can safely go back to the old directory and everything still works. The added benefit of this is with rollbacks we won't touch the /usr/local/bin symlink so it should still work.

Test Plan: upgrade from failing 2024.1.3 to 2024.2 with this applied and ensure rollback still works.

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39170