[BACKPORT 2024.2][#24157] docdb: Update missing_value for clone in import snapshot
Summary:
Original commit: 69103989f51c65f7a86ab1cdcea5b28b7b2bfdfd / D38971
Since https://phorge.dev.yugabyte.com/D25297, the missing_value field of a tablet is
populated on ADD COLUMN with a non-volatile default value.

missing_value is not set by ysql_dump. Backup/restore handles this by updating the missing value
from the tablet metadata (in `RestoreMissingValuesAndMergeTableSchemaPackings`). This does not work
for clone because we can clone to a time before dropping a column with a default value, in which
case missing_value is also dropped in the tablet metadata.

The correct value of the field is stored in the master schema (which clone reads as of a
point-in-time), so in this diff I pass that value to the tablet via the AlterTable in ImportSnapshot.

As the alter command that populates the missing_value is asynchronous, there might be a window for clone where we can't read the missing value (and get a query error). We should eventually either wait on these alters, or move all the alter logic into RestoreSnapshot.

Jira: DB-13044

Test Plan:
Enabled the following non-colocated and colocated (respectively) tests:
```
build release --cxx-test yb-admin-snapshot-schedule-test --gtest_filter *PgsqlDeleteColumnWithMissingDefault/3 --with-tests
build release --cxx-test yb-admin-snapshot-schedule-test --gtest_filter *PgsqlDeleteColumnWithMissingDefault/4 --with-tests
```

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39181