[BACKPORT 2024.2][PLAT-15675] Show tables in error state with higher priority in UI
Summary:
Original commit:
9e07477221320aa7b5917b47fbaf3c1d038553b2 / D39435

Before the tables would be shown based on the enum order for the statuses of the `XClusterTableConfig`. However this is incorrect as we should sort by `code` order which is similar to severity.

Test Plan:
Add unit test

Also, create dr config and then add table to only source universe. Make sure the extra table on source appears first
{F304133}

Reviewers: sanketh, hzare, jmak, vbansal

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39436