[#24501] YSQL: Upgrade template1 database in major version upgrade
Summary:
Yugabyte hosts shared catalog tables such as pg_database in namespace template1. This requires special handling for upgrade, because the upgrade drops and re-creates all built-in databases in order to preserve their characteristics, but the shared tables need to be kept around.
Jira: DB-13418

Test Plan:
Jenkins

On MacOS arm64:
  ./yb_build.sh release --sj --skip_tests
  ./yb_build.sh release --cxx-test pg15_upgrade-test --with_tests
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

On AlmaLinux 8:
  ./yb_build.sh fastdebug --gcc11 --sj --skip_tests
  ./yb_build.sh fastdebug --gcc11 --cxx-test pg15_upgrade-test --with_tests
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

Reviewers: hsunder, fizaa

Reviewed By: hsunder

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D39171