YSQL: Fix for test failures introduced in D38246
Summary: The diff D38246 introduces a new flag that makes superuser connections using Connection Manager sticky by default. This change is not meant to be used for the java test framework, and was not disabled for Connection Manager tests. This patch is a quick fix for the same.

Test Plan:
Jenkins: all tests

Verify using detective that the tests failing are in the vicinity of (newly) expected failures on master after having some tests being fixed by D38246

Reviewers: stiwary

Reviewed By: stiwary

Differential Revision: https://phorge.dev.yugabyte.com/D39244