[#24531] YSQL: conn-mgr: Fix the TestPgBatch#testSchemaMismatchRetry unit test failure
Summary:
In the test **TestPgBatch#testSchemaMismatchRetry**, the test will not work as expected when connection manager is enabled. The change is to run the queries in a ROUND_ROBIN mode and to make sure the test scenario is the same with connection manager and without connection manager. In this case, we are making the query `UPDATE t SET v=2 WHERE k=0 `and the batch queries to run on the same backend.
Jira: DB-13565

Test Plan:
Jenkins: test regex: .*TestPgBatch.*, enable connection manager

Check that UT is passing now with connection manager enabled and disabled.
```
./yb_build.sh --enable-ysql-conn-mgr-test --java-test 'org.yb.pgsql.TestPgBatch#testSchemaMismatchRetry'
```

Without dd60793542b3, the test fails as expected, (tested with and without --enable-ysql-conn-mgr-test)

```
ts1|pid732863|:10881|http://127.25.214.15:21780 2024-10-21 16:44:46.320 UTC [733471] LOG:  execute <unnamed>: UPDATE t SET v=2 WHERE k=1
16:44:46.360 (Time-limited test) [ERROR - org.yb.AssertionWrappers.logError(AssertionWrappers.java:41)] Assertion error:
java.lang.AssertionError: Internal retries are not supported in batched execution mode: expected java.sql.BatchUpdateException to be thrown, but nothing was thrown
        at org.junit.Assert.assertThrows(Assert.java:1028)
        at org.yb.AssertionWrappers.lambda$assertThrows$53(AssertionWrappers.java:831)
        at org.yb.AssertionWrappers.wrapAssertion(AssertionWrappers.java:46)
        at org.yb.AssertionWrappers.assertThrows(AssertionWrappers.java:830)
```

Reviewers: mkumar, rbarigidad, devansh.saxena, skumar, stiwary, patnaik.balivada

Reviewed By: patnaik.balivada

Differential Revision: https://phorge.dev.yugabyte.com/D39251