[PLAT-15780] yba-ctl install failing on CIS hardened image
Summary: We have permission issues on CIS hardened image where umask is 077 where certain files like node-agent and ybc are not owned by yugabyte. This causes issues with YBA start when we try to move files and don't have the permissions necessary. This diff is more aggressive about changing the permissions at the top level of all commands before we start the processes to make sure the data and software directories that we created are completely owned by the service username.

Test Plan: install on umask with 077, ensure no files in /opt/yugabyte are owned by root

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39256