Merged PR 1490191: [OSS] Fix compiler error seen in Ubuntu 24.04
### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Fix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No
### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
Fix compiler error - looks like that compiler seems to be a bit aggressive on uninitialize - the code is correct but fails to compile on 24-04 and PG16. Fix the initialization.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to address a compiler error in Ubuntu 24.04.

#### PR Summary
This pull request fixes a compiler error encountered in Ubuntu 24.04 by making necessary code adjustments.
- Modified `oss/Makefile.cflags` to remove an unnecessary compiler flag.
- Updated `oss/pg_helio_api/src/aggregation/bson_aggregation_window_operators.c` to initialize `nConst` to `NULL` before usage.