[BACKPORT 2024.2][PLAT-15733] Retrying asRoot migration
Summary:
Because of the issue described in D39166 we were not performing asRoot migration. It is safe to run again so we will simply retry it with higher key.
Original commit: 13ee5d743e5c014c9ce67af0db2b36595fa12542 / D39167

Test Plan: apply on top of upgrade from 2024.1.3 to make sure it fixes the problem

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39268