[PLAT-15803]make id token groups claim runtime config
Summary:
We fetch the list of groups from id token using the "groups" claim.
Make this field runtime config since this configurable on the server side.
Will also add a ticket to add a UI field to configure this on the OIDC config page.

Test Plan: Existing UTs cover the change.

Reviewers: #yba-api-review!, svarshney

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39294