[#24439] YSQL, QueryDiagnostics: Add function for canceling query diagnostics
Summary:
This diff adds `yb_cancel_query_diagnostics()`, which can be used to cancel an ongoing query diagnostics bundle. This can be especially useful in cases where the query diagnostics bundle is initiated unintentionally or the diagnostics_interval_sec is set longer than intended, and a cancellation mechanism is required to halt the process.
The function must be supplied with the `query_id` to remove the entry from being bundled.
This diff adds `DIAGNOSTICS_CANCELLED` state for the cancelled bundles.
Jira: DB-13346

Test Plan: ./yb_build.sh --java-test TestYsqlUpgrade

Reviewers: asaha, telgersma

Reviewed By: asaha, telgersma

Subscribers: telgersma, yql

Differential Revision: https://phorge.dev.yugabyte.com/D39041