[BACKPORT 2.20][PLAT-13436] Updated access key in redux store when modifying providers
Summary:
The create/edit universe form takes the cached access key data from redux store when determining the list of
access keys to show on the form for a given provider.
As the universe form itself does not update the access key list, it may be stale if the access key was updated.
Reloading the webpage will trigger a refetch from `AuthenticatedComponent`, but it would be good to update the stored
values when we make changes to the provider.

This diff adds logic for updating these cached access key values in redux.
In the future, we'll be removing all these dependencies on cached data in redux and centralizing it to
just React Query cache. That will help decrease the bugs due to stale data being read.

Original commit: 516b9e3fed08767b15f3cdad02398fb8655fae85 / D39327

Test Plan:
- Create an onprem provider with an ssh key named with invalid characters
- Observe that the provider creation fails and no access key is shown on the provider.
- Edit the provider and upload the access key with a valid name this time.
- Go to the create universe page and verify that the access key is available under our new onprem
  provider.
{F303049}

Reviewers: rmadhavan, lsangappa, kkannan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40362